chore: move pages under shared resources menu header
